requirements_production.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
gunicorn==21.2.0
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6
aiohttp==3.9.1
